Send notification to Pebble on new message
This implements basic notifications to the Pebble through the app (using
an intent).  This simply hooks into NotificationService.notify().

This is pretty basic, but it works (I haven't tested to see how the
intent is received when the Pebble app is not around, though). More
fancy stuff could probably be added to avoid getting flooded, but the
Pebble app already does a good job a filtering notification (e.g.,
screen on or quiet times).

Signed-off-by: Olivier Mehani <shtrom@ssji.net>