disable call integration on Android 15
The call integration depends on the operating system giving us a list of
possible output devices (speaker, earpiece, BT headset, ...) via a
callback called
`onAvailableCallEndpointsChanged()` that is called as soon as the app
registers the call with the operating system.
For outgoing calls on Android 15 this method is no longer invoked
leaving Conversations without knowledge about what devices are available
and thus without the ability to switch between speaker and earpiece.

Logcat output seems to indicate that there might be some sort of race
condition (there is a line that says something like: available endpoints
changed but service was not bound)
There is further evidence in that if one tries multiple calls in a row
(10+) one might actually see the output device switcher / indicator.

However after a full day of investigation it seems that the race
condition is in the platform code; not in Conversations.

(cherry picked from commit be0b5f84d7cd0af7a8dab7f433820255c7cdc804)