fix: second reply treated as separate message
diagnosis: when we get a message with a <replace> element,
we overwrite several fields of the replaced message. Unfortunately, this
includes the remoteId of the replaced message, which is the field by
which we look up the message to be replaced. All corrections to a
message reference the id of the original, unedited message, so edits
>= 2 find a null replacedMessage.

fix is just to remove that line

(cherry picked from commit 7522d7eab5d8e7b48031f01f91267c6f4234bb9a)